requirements.txt

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
gradio==4.44.0
firebase-admin==7.1.0
httpx[http2]==0.28.1
requests==2.32.3
aiokafka==0.10.0
websockets==12.0
pytest==8.3.3
